git: fix build errors
It appears that square brackets are troublesome, because they confuse the
parser into thinking an argument to \compendium has been passed, or something.
I'm not sure. Surrounding them with "{[}" seems to help (suggestion from
http://latex.org/forum/viewtopic.php?t=10934).

\label doesn't seem to work; not sure why, but I don't think we care about it,
so I just punted it.

I also removed a ton of the placeholder SVN content; I'm not sure that matters,
but it's easy enough to copy from guildcamp-svn, rather than editing the
bottom, and I think the content will be pretty different.

Note also that deleting *.aux and/or *.out seems required to fix a broken
build, for some reason.