GitLab: set up pages
So far, this is a Pages setup that's reasonable for a game to use --
presumably the rules & scenario is going to be public -- and a good
start for a game to add more content (eg, public greensheets).
Presumably in that case index.html's content would move to a file,
rather than being generated inline in gitlab-ci.yml.

As part of this, I've also made the Production Makefile build the rules
doc, which seems fairly reasonable. If a bunch more handouts were added,
that directory should probably get its own Makefile.